Do not wrap particles until rebuilding the tree
This was causing the tree to never be rebuilt because the particles were
wrapped back into the primary image before the displacement check was
performed. And since the diaplacement check used the minimum
plane-to-plane distance, the particle could never really move far enough
to trigger a rebuild. This may have been leading to rare overlaps.